Add call to eager_autoload!
During testing on 5.x versions of Logstash, I encountered some weird
dependency errors, which appear to be related to the issues
described in https://aws.amazon.com/blogs/developer/threading-with-the-aws-sdk-for-ruby/
which can occur in versions of Jruby predating jruby-9k.

This commit adds the call to eager_autoload! as recommended in the
article above

Fixes #4